refactor(workspace): rename create-document → create-documents to match export
The file exports createDocuments (plural) but was named create-document
(singular). Every other file in workspace/ follows the convention:
create-table.ts → createTable, create-tables.ts → createTables,
create-workspace.ts → createWorkspace. Fix the only outlier.
